requirements.txt
asgiref==3.6.0
boto3==1.26.113
botocore==1.29.113
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.7
django-cors-headers==3.14.0
django-rest-passwordreset==1.3.0
django-storages==1.13.2
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
packaging==23.0
Pillow==9.4.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
requests==2.28.2
ruamel.yaml==0.17.21
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
whitenoise==6.4.0
